modified:   Windows/Lenovo/LSUClient.ps1 	modified:   Windows/Lenovo/LSUClient_auto.ps1

+++ Windows/Lenovo/LSUClient.ps1
@@ -71,8 +71,13 @@ for ($Round = 1; $Round -le $MaxRounds; $Round++) {
 }
 
 # Cleanup Files
-Write-Host "Cleaning up..."
+Write-Output "Cleaning up..."
+if (Test-Path -Path $env:TEMP\LSUPackages) {
 Remove-Item -Path $env:TEMP\LSUPackages -Recurse
+}
+else {
+    Write-Host "Nothing to clean up"
+}
 
 # Prompt for reboot
 Restart-Computer -Confirm:$true
